Consistency on SagaConsumeContext, ensuring that CorrelationId is always Saga.CorrelationId, so that InitiatorId matches the saga CorrelationId on published/sent messages.
