+++ src/MassTransit/Saga/Factories/DefaultSagaFactory.cs
@@ -31,7 +31,7 @@ namespace MassTransit.Saga.Factories
 
             var instance = SagaMetadataCache<TSaga>.FactoryMethod(context.CorrelationId.Value);
 
-            var proxy = new NewSagaConsumeContext<TSaga, TMessage>(context, instance);
+            var proxy = new DefaultSagaConsumeContext<TSaga, TMessage>(context, instance);
 
             proxy.LogCreated();
